+++ browser/render/layouts/list-and-detail/snippet-detail-multi-file/index.jsx
@@ -210,7 +210,10 @@ export default class SnippetDetailMultiFile extends React.Component {
     // if the mode for that language exists then use it otherwise use text
     if (resultMode) {
       const snippetMode = resultMode.mode
-      if (snippetMode === 'htmlmixed') {
+      if (snippetMode === 'null') {
+        editor.setOption('mode', 'null')
+        editor.setOption('htmlMode', false)
+      } else if (snippetMode === 'htmlmixed') {
         require(`codemirror/mode/xml/xml`)
         editor.setOption('mode', 'xml')
         editor.setOption('htmlMode', true)
@@ -293,7 +296,10 @@ export default class SnippetDetailMultiFile extends React.Component {
         // if the mode for that language exists then use it otherwise use text
         if (resultMode) {
           const snippetMode = resultMode.mode
-          if (snippetMode === 'htmlmixed') {
+          if (snippetMode === 'null') {
+            editor.setOption('mode', 'null')
+            editor.setOption('htmlMode', false)
+          } else if (snippetMode === 'htmlmixed') {
             require(`codemirror/mode/xml/xml`)
             editor.setOption('mode', 'xml')
             editor.setOption('htmlMode', true)
@@ -476,6 +482,11 @@ export default class SnippetDetailMultiFile extends React.Component {
   handleSnippetLangChange () {
     const { editor } = this.refs
     const snippetMode = CodeMirror.findModeByName(this.refs.lang.value).mode
+    if (snippetMode === 'null') {
+      editor.setOption('mode', 'null')
+      editor.setOption('htmlMode', false)
+      return
+    }
     require(`codemirror/mode/${snippetMode}/${snippetMode}`)
     editor.setOption('mode', snippetMode)
   }